backend/requirements.txt
flask
pillow
amulet_nbt
